Constructor for a Cost-Effectiveness Tree Object

The resulting object is used in the main package functions. Branch distributions are currently permitted as uniform, gamma and triangle. All branches require some distribution. All branches must end in a terminal node.

costeffectiveness_tree(yaml_tree, details = "", data_prob = NA,
  data_val = NA, ...)

Arguments

yaml_tree
YAML file or location address

details
General details of decision tree

data_prob
Branching probability data

data_val
Cost or health data

...
Additional arguments to be passed

Value

list of class costeffectiveness_object

See also

calc.expectedValues

Examples

CEtree <- costeffectiveness_tree(yaml_tree = "raw data/LTBI_dtree-cost-distns.yaml")

#> Warning: cannot open file 'raw data/LTBI_dtree-cost-distns.yaml': No such file or directory

#> Error in readLines(con): cannot open the connection

print(CEtree)

#> Error in print(CEtree): object 'CEtree' not found

CEtree$osNode <- calc.expectedValues(CEtree$osNode)

#> Error in calc.expectedValues(CEtree$osNode): could not find function "calc.expectedValues"

print(CEtree)

#> Error in print(CEtree): object 'CEtree' not found